requirements.txt
transformers
sentence-transformers
faiss-cpu
torch
numpy
pandas
streamlit
flask
datasets
detoxify

